backend/requirements.txt
# Django and core dependencies
Django>=5.2.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS handling
django-cors-headers>=4.3.1

# Environment variables
python-dotenv>=1.0.0
dj-database-url>=2.1.0

# HTTP requests (for Paystack API integration)
requests>=2.31.0

# Additional utilities
Pillow>=10.0.0  # For image handling
img2pdf>=0.5.1  # For converting images to PDF

# Cloudinary integration
cloudinary>=1.40.0
django-cloudinary-storage>=0.3.0

# Production server
gunicorn>=21.2.0

# Static files serving
whitenoise>=6.6.0

# Content Security Policy
django-csp>=3.7

# Database drivers and connection pooling
PyMySQL>=1.1.0  # MySQL driver
psycopg2-binary>=2.9.7  # PostgreSQL driver (for production)

# Caching and performance
redis>=4.6.0  # Redis client
django-redis>=5.4.0  # Django Redis integration

# Additional performance dependencies
psutil>=5.9.0  # System monitoring

# Background task processing
celery>=5.3.0  # Background task queue
django-celery-beat>=2.5.0  # Periodic tasks
django-celery-results>=2.5.0  # Task results backend

